Update code formatting and enhance Makefile
The code formatting has been updated to adhere to conventions and improve the readability of various Swift scripts. Additions have been made to the Makefile to support the 'format' target, allowing for recursive application of formatting rules in parallel, where practicable. The changes also included safeguard against attempting formatting in the case where 'swift-format' is not available.