Add versioning from git tags to app
The modifications in this commit include changes in the "Package.swift", "PanelPulley.swift", "Makefile", and ".gitignore" files to incorporate versioning from git tags. The implementation involves writing the current version into a "version.txt" file during the build process in the "Makefile", and retrieving this version to use in the application in the "PanelPulley.swift" file. The ".gitignore" file is updated to exclude the "version.txt".